requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.4
markdown-it-py>=3.0.0
httpx>=0.28.1
deepagents>=0.4.12
langchain>=1.2.15
langgraph>=1.1.6
langchain-openai>=1.0.0
pydantic>=2.11.0
pydantic-settings>=2.10.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
typer>=0.16.0
rich>=14.0.0
pytest>=8.4.0
pytest-asyncio>=0.26.0

